feat: retire-workers-test-only-env-slice-parser-001 - Move the shared env slice parser to the workers test surface
